Move generic_log_worker/src/ctlog.rs to log_ops.rs
