+++ crates/generic_log_worker/src/log_ops.rs
